Document Ingestion/requirements.txt
openai==1.55.3
Flask==2.3.2
azure-identity==1.16.1
pymongo==4.7.0
flask-swagger-ui==4.11.1
opencensus==0.11.4
opencensus-ext-flask==0.8.2
opencensus-ext-azure==1.1.13
Werkzeug==3.0.3
Flask-CORS==5.0.0
pycryptodome==3.21.0
pdfplumber==0.10.2
PyPDF2==3.0.1
python-dotenv==1.0.1
azure-storage-blob==12.19.0 
azure-servicebus==7.12.1
aiohttp==3.9.5
motor==3.5.1
pdfservices-sdk==4.0.0
pymupdf==1.24.13
reportlab==4.2.5
pandas==2.2.3
openpyxl==3.1.4
urllib3==2.2.2
pillow==10.3.0
zipp==3.19.1
dnspython==2.6.1
flask-talisman==1.1.0